try this

+++ clean_data.py
@@ -7,10 +7,14 @@ from wurlitzer import pipes
 
 fnames = glob.glob("**/*.fit*", recursive=True)
 for fname in tqdm.tqdm(fnames, desc="removing fits files"):
+    if "data_final" in fname or "mdet_data" in fname:
+        continue
     os.system("rm -f " + fname)
 
 fnames = glob.glob("**/*.hs", recursive=True)
 for fname in tqdm.tqdm(fnames, desc="removing healsparse files"):
+    if "data_final" in fname or "mdet_data" in fname:
+        continue
     os.system("rm -f " + fname)
 
 fnames = glob.glob("**/*.ipynb", recursive=True)
